Enhance CI workflow for Windows: add artifact compression and release steps; refactor config and profile tests, and rename validation functions for consistency.
