+++ .github/workflows/go.yml
@@ -3,11 +3,11 @@ name: Go
 on:
   push:
     branches: [ "main" ]
+    tags: [ "v*" ]
   pull_request:
     branches: [ "main" ]
 
 jobs:
-
   build:
     runs-on: windows-latest
     steps:
@@ -19,7 +19,61 @@ jobs:
         go-version: '1.23'
 
     - name: Build
-      run: go build -v ./...
+      run: go build -v -o PowerShellProfileLauncher.exe ./...
+
+    - name: Compress Artifact to Zip
+      run: Compress-Archive -Path .\PowerShellProfileLauncher.exe, .\profiles.csv.tmpl, .\config.json.tmpl -DestinationPath .\PowerShellProfileLauncher.zip
+
+    - name: Upload Build Artifact
+      uses: actions/upload-artifact@v3
+      with:
+        name: PowerShellProfileLauncher
+        path: PowerShellProfileLauncher.zip
+
+  test:
+    runs-on: windows-latest
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Set up Go
+      uses: actions/setup-go@v4
+      with:
+        go-version: '1.23'
 
     - name: Test
       run: go test -v ./...
+
+    - name: Super-Linter
+      uses: super-linter/super-linter@v7.2.0
+      with:
+        github_token: ${{ secrets.RELEASER }}
+        validate_json: true
+        validate_go: true
+         
+
+  release:
+    runs-on: windows-latest
+    needs: [build, test]
+    if: startsWith(github.ref, 'refs/tags/')
+    steps:
+    - uses: actions/checkout@v4
+
+    - name: Download Build Artifact
+      uses: actions/download-artifact@v3
+      with:
+        name: PowerShellProfileLauncher
+
+    - id: create_release
+      name: Create Release
+      uses: ncipollo/release-action@v1.14.0
+      with:
+        name: ${{ github.ref_name }}
+        token: ${{ secrets.RELEASER }}
+        artifactErrorsFailBuild: true
+        discussionCategory: 'Releases'
+        generateReleaseNotes: true
+        makeLatest: true
+        skipIfReleaseExists: true
+        tag: ${{ github.ref_name }}
+        body: ${{ github.event.release.body }}
+        artifacts: PowerShellProfileLauncher.zip
